Enable gradle caching
This commit will enable gradle caching. Enabling caching means that
tests only run for a module if there are changes in that module or a
dependency. This can speed up builds for many small changes and makes
reruns because of a single test failure much faster, but does not give
us as much insight into flaky tests or potentially hidden dependencies.

Also, change the codeCoverageReport to always run since this task does
not explicitly declare its dependencies (it creates a report for
whatever test cases were run in the build but not necessarily all test
cases).