[StatsMigration] Remove database rows for missing partition/account/container in AccountReports  (#1822)
When a new StatsWrapper is generated and persisted to database, we only override the database rows where the storage value is different. Since we have a copy of previous StatsWrapper, we can easily know which database rows to update.
However, we don't remove those database rows when any partition or account or container no longer exists in the host, we just silently ignore it.

This PR would take the absence of partition/account/container into consideration and remove corresponding database rows. We are not using batch operations for these DELETEs, since we know the delete should be a rare operation.