Support async get valid data size (#1852)
While debugging the dotted metric issue, we found that one piece of code using gauge takes a really long time(5 to 10 s) to get the size of valid data at a particular point in time for all log segments. And the reason this issue only happens in the main cluster is we have many small blobs and many partitions in the main cluster are almost full. This pr support to calculate the valid data size in a background thread that won't block metrics collection.