Add a gauge metric to record the number of log segment compacted in compaction (#1865)
Adding a metric to record how many log segments we can save from each round of compaction.
We are expecting at least 1 log segment saved in StatsBasedCompactionStrategy.

This PR introduced a complex way to count how many segments are compacted because this PR also handles the case where process failed in the middle of compaction. The easy way is to have a variable to store the number of log segment before compaction and then after compaction we just subtract this number. But this won't work when the process fails in the middle of the compaction. That's why we need this complex method.