Make the data size for request quota calculation configurable. (#1913)
Until now we used 4MB chunks as unit for quota calculation. From experiments we have seen that this is much larger and users have a significant number of blobs with smaller blob size. So having a configurable unit gives us flexibility to experiment with smaller chunks to see what makes sense in different clusters.