Server certificate validation (#1952)
Verify server identity on the router via Subject Alternative Names part of the certificate if the newly introduced ambry.http2PeerCertificateSanRegex parameter is set.

If the parameter is set, server tries to match at least one of the subject alternative names to the corresponding regular expression. In the case of failure, router closes the channel established between the server.