Use AdaptiveOperationTracker in PutOperation (#1937)
We have some issues with SimpleOperationTracker with PutOperation. We want to change PutRequestParallelism to 2 and expect that only 2 copies of the same chunk will be sent out to ambry-server replicas. However, this is not what we see from the metrics. Even with parallelism changed to 2, we are still seeing 3 copies of the same chunk sent out to replicas. The reason of this is because of SimpleOperationTracker.

In SimpleOperationTracker, when we send out first two PutRequests, it will block because of the parallelism is 2. But as long as we receive the response from any of these 2 requests, the inflightCount would be decremented, thus we will immediately send out the third request. For that, we end up with 3 requests.

One way to fix this is to set the PutRequestParallelism to 1, in this case, we will send the first request out, wait and receive the first response. Then send the second request out and receive the second response. Assume that both responses are successes, we would stop and acknowledge this PutOperation as successful. But this way, we would drastically increase the latency of PutOperations, thus damage the overall upload performance.

The other way is to have different logic in the OperationTracker, to send out first 2 requests and wait until both responses come back. But this way, if we have a slow replica, then we are damaging the performance as well.

The best solution is to use AdaptiveOperationTracker, so that when we have slow replica, we would send out the third request right away. But before we can use AdaptiveOperationTracker, we have to change some logic in it.

AdaptiveOperationTracker has the same issue as the SimpleOperationTracker, where when the PutRequestParallelism is 2, it will send out the third request as soon as it receives the first response. This PR would fix that logic.