Add addl azure auth and key storage options
Provide a way to fetch a Cosmos DB from an azure keyvault protected by
RBAC instead of directly storing the key in configs.

Provide a StorageClient impl that uses the azure-identity library. This
library internally handles token refresh and may be preferable to the
lower level msal4j library after we have verified it.