clean up chunks in put manager (#2025)
There is a bug in the PutManager that could cause memory leak.
When the client is uploading a blob and the connection is terminated, we would set the operation to be completed in the callback. However, we only release the ready and completed chunks in the callback method, not the building chunk. We are expecting the chunk filler thread would release the building chunk, however, the chunk filler thread might be sleeping and this operation would be removed by put manager. In this case, the building chunk is never released.

This PR would fix this issue by releasing all the chunks in PutManager when the operation is removed.