Surround NonBlockingRouterTests with try-catch block to make sure cleanup happens for each test, and one test failure doesn't cause subsequent tests to fail. Also change timeout to wait for test operations to complete to 15 secs from 10 secs. (#2033)
Surround NonBlockingRouterTests with try-catch block to make sure cleanup happens for each test, 
and one test failure doesn't cause subsequent tests to fail. Also change timeout to wait for test 
operations to complete to 15 secs from 10 secs.

NOTE: The only change in this file is to surround the tests that attempt to close router at the end, 
with try-catch block. There are no logic changes in this file. Note that it might be possible to unify
the cleanup logic by refactoring the code a little. But this change keeps that out of its scope.