Aggregate Get and Put qps to detect live traffic before container deletion (#2071)
We plan to automatically test live traffic before the container deletion.
In order to do that, Nuage team support a flow which can only take two metrics, one for put request and one for 
get request.
So this pr support

    Aggregate the per request QPS metric into the get and put QPS metric.
    Code refactoring to move containerMetricsEnabledRequestTypes into open source, so we won't forget to define 
the request type in close source. Currently we forget to include UndeleteBlob and PutBlob into the list in close source for container metrics.