Handle quota rejected response in all operations. (#2086)
This PR handles quota rejected response in all operations -- Get, GetInfo, Put, Delete, Undelete and TtlUpdate.
The main changes are in the handleResponse() of each operation. In handleResponse() if the response info indicates that the response has been rejected due to quota compliance, then the operation/chunk should be marked as failed due to RouterErrorCode.TooManyRequests.

This PR also updates the precedence of errors for each of the operations, by updating the getPrecendenceLevel() method in each of the operations. In case multiple errors are present, the final error thrown is based on the precedence order.