Pass blobId to Azure StorageClient downloadWithResponse function (#2099)
In the big picture, StorageClient will be an implementation of an interface called AzureStorageClient.

                   AzureStorageClient
                      /                    \
         StorageClient      ShardedStorageClient
AzureStorageClient will expose the same interface that StorageClient does today. Interaction with a particular storage account will be contained in StorageClient whereas ShardedStorageClient will pick the StorageClient instance corresponding to a blob. This is a temporary scheme to be backward compatible, i.e. we can go back and forth between the sharded and unsharded implementation until we can fully switch to the sharded strategy. Once that happens, we will dismantle the AzureStorageClient interface and StorageClient will be private to ShardedStorageClient.

This is in preparation to pick the storage account corresponding to a blob upon sharding the blobs across multiple storage accounts by the ShardedStorageClient. PR passes the blobId to the Azure StorageClient downloadWithResponse function when applicable (this function is also called to fetch metadata files regarding compaction and replication) so that future PRs can extract the partitionId from the blobId and map the request to the correct storage account.

PR also makes the testConnectivity and deleteFile routines public since these routines will be called through the AzureStorageClient interface.