Fix a potential memory leak in http2networkclient (#2116)
In drop request, we only retrieve the request info from the channel attribute map, which means we always keep a reference of request info in the attribute map. If this request info somehow has a memory leak, we will never detect it since there is always a valid reference.

This should invalidate the reference of dropped requests in attribute map. Hopefully this would fix the memory spike issue or at leasts show more information in the log files.