[Frontend] Add metadata cache in router
This patch adds a cache to save metadata chunks of composite blobs.
The NonBlockingRouter instantiates a volatile MetadataCache. The cache
maps a String blobID to a MetadataChunk object. We use Guava Cache and
configure it to hold a maximum of routerMetadataCacheMaxItemCount items.
Localhost experiments show we save 110-125ms on RANGE requests on
a 1TB blob with range sizes varying from 1MB to 1024MB.