Temporarily return 404 when two replicas not found but one succeed (#2217)
There’s a case that due to some reason two replicas don’t have the data(maybe due to OSUA wiping out some data), but the third one still has the DELETE records. In current scenario, we return 503 which is not appropriate.

This is the short term fix to revert the changes to the previous version which return 404. For the long term fix, we should check all replica status for delete operation and return succeed(200) when one replica succeed.