Remove some unused BlobStoreCompactor configuration (#2244)
In this PR, we remove two unused configurations for BlobStoreCompactor.

store.always.enable.target.index.duplicate.checking: This configuration was used to always check if there is a duplicate in target index in compaction. We don't have to do this since only the crashed compaction would need to check duplicates in target index.
store.compaction.enable.basic.info.on.missing.duplicate: This configuration is used to print out some basic index and entry information when we are expecting duplicates and not getting any. Change it to trace log.