Update Hard Delete to support Blob Format V3. (#2274)
* Update Hard Delete to support Blob Format V3.

HardDelete was originally written in 2016.  It zeros user metadata and blob data.  I believe it is deprecated and replaced by compaction. Not sure whether it is still active, but the code is there, so update it.

* Update Hard Delete to support Blob Format V3.

HardDelete was originally written in 2016.  It zeros user metadata and blob data.  I believe it is deprecated and replaced by compaction. Not sure whether it is still active, but the code is there, so update it.

* Update Hard Delete to support Blob Format V3.

HardDelete was originally written in 2016.  It zeros user metadata and blob data.  I believe it is deprecated and replaced by compaction. Not sure whether it is still active, but the code is there, so update it.

* Update Hard Delete to support Blob Format V3.

HardDelete was originally written in 2016.  It zeros user metadata and blob data.  I believe it is deprecated and replaced by compaction. Not sure whether it is still active, but the code is there, so update it.