Compression library accept ByteBuffer.
Previous implementation accepts byte[] or Heap.  It relies on JVM pinning to avoid memory copy when calling native compression library.  However, it is not reliable and unpredictable.  A better alternative is to use ByteBuffer that can hold either Heap or Direct memory.  This change adds ByteBuffer to API.  After ByteBuffer API deployed, the byte[] will be deleted.