Replication Optimization: Treating deleted put record's size same as delete  (#2297)
This is an optimization for replication.
In replication, when replicator sends a ReplicaMetadataRequest to remote server, remote server would scan through the index segment files. It will stop when the number of scanned bytes are larger than the max number of byte in the ReplicaMetadataRequest. And the max number of bytes is usually 4MB (one PutBlob size).

In this case, when the remote server sees a Put record in the index segment file, it will stop scanning right away. However, if this Put record is already deleted (there is a Delete record after this Put), the size in the MessageInfo would be changed to size of Delete. When the replicator sees this id, it knows that this id is deleted in the remote server so it does nothing. This is a terrible waste of roundtrip since we are spending one round trip to deal with one blob id.

In this PR, we would keep scanning if this put Record is already deleted. In this way, we can return more ids back to the replicator in one request. We only do this for IndexBased replication token.