Update logic for sending 404 and 503 in SimpleOperationTracker (#2333)
This PR refactors SimpleOperationTracker code to use originating data center as the source of truth when sending 404 responses.

If all the replicas in originating data center are up and return that blob is not present, we send 404. Config routerOperationTrackerTerminateOnNotFoundEnabled is used to terminate the operation immediately.

If any of the originating DC replicas are not up (i.e. in leader or standby state) and the success count doesn't meet the quorum, we send 503 if there are no other errors with higher precedence (such as Authorization failure, Blob deleted, etc).