Update clustermap and state transition factory to deal with duplicate partition ids (#2336)
Ambry is currently going through the process of reconstructing resource names. We used to assign partition ids to several resources, even if those partition ids belong to the same clique. Now we are trying to gather all the partition ids of the same clique to one resource.

Since in server we honor new resource over the old resource, we have to do the same thing in frontend. This is easy since we can compare the resource names and know that higher number resource name would win.