Add configuration to enable nonblocking network client in local colo replication (#2371)
This PR adds a configuration to add nonblocking network client for intra-colo replication.

Intra-colo replication is much faster than inter-colo replication since it has less data to replicate and faster connection to replicate data with. That's why a blocking channel usually works just fine for intra-colo replication. But this change would still be beneficial for intra-colo replication especially when it comes to bootstrapping.