[FULL_AUTO] ClusterMap: Add a method to return if a data node is on FULL_AUTO (#2370)
This PR adds a method in clustermap interface and HelixClusterManager implementation to indicate a data node is on FULL_AUTO or not by returning true or false.

When turning on FULL_AUTO, we would first reconstruct resource so that a data node can have partitions from only one resource. Then we update the resource's ideal state to change the RebalanceMode to FULL_AUTO. In this way, a data node would know if all the partitions assigned to it are on FULL_AUTO mode or not.