Reducing the Number of StoreExceptions for checking MountPath
The disk healthchecker will check if the mount is accessible every
180s, so if a mount is inaccessible, the logs will be crowded with
repeated exceptions of disks being inaccesssible. With this change, we reduce the
number of exceptions. So, we only throw an exception, if the previous state wasn't
MOUNT_NOT_ACCESSIBLE.The initial state for diskHealthStatus will be null
to trigger that initial Store Exception, if the mount isn't accessible.