Reducing the Number of StoreExceptions for checking mount path (#2381)
The disk healthchecker will check if the mount is accessible every 60s, so if a mount is inaccessible, the logs will be crowded with repeated exceptions of disks being inaccessible. With this change, we reduce the number of exceptions. So, we only throw an exception, if the previous state wasn't MOUNT_NOT_ACCESSIBLE.The initial state for diskHealthStatus will be null to trigger that initial Store Exception, if the mount isn't accessible. Build was successful