[DR] Use static map to discover VCR nodes
It's becoming difficult to debug VCR issues because the current
setup uses HelixVcrCluster to dynamically discover vcr nodes and
then randomly picks a node to talk to. It is also difficult to play
around with VCR configs and understand what works because one has to
change configs in all VCR nodes and we don't know which VCR node will
picked up next time we start a DR node.

This patch adds a support for reading files to discover vcr nodes and
allows us to fix a vcr node for an experiment or debugging.