Adding a new configuration for fetch size in bootstrap mode (#2389)
When the host is going bootstrap, we should increase the fetch size to make bootstrap faster. There are different ways to make this happen, but the easiest way is to define a new configuration and assign a bigger value to this configuration.

Later, we can have a smarter ways of increase the fetch size by dynamically increase it until we see damaged network performance.