Reserve metadata id for signed-url request for POST with chunked-upload enabled. (#2449)
There are 4 cases to handle in order to reserve metadata chunks:

Reserve metadata chunk id for regular Post requests to upload a regular simple or composite blob.
Reserve metadata chunk id for POST signed-url requests with the "chunked-upload" header set to true. This reserved metadata id will be used for chunked uploads and the subsequent stitch blob operation.
For chunked uploads use the reserved metadata id encoded in the signed URL and after upload encode the reserved metadata id in the signed-id of the chunk.
For stitch requests use the reserved metadata id in the signed-ids of the chunks.
This PR handles case 2.