Add a network client implementation to communicate with cloud (#2470)
Adding a NetworkClient implementation to return responses by talking to cloud. This would be useful in RecoveryManager to recover data from cloud, with a nonblocking network client. However, in this implementation, the client is a blocking client. But it's fine to be in this way. If later we can have a real nonblocking implementation, we can later replace it.