Add changes to make disk selection locally when rolling back from Full-Auto to Semi-auto (#2504)
During roll back to Semi-auto, we intend to use the replica placement calculated by Helix (in Full-auto mode) to avoid too many replica movements. However, this means that when bootstrapping a replica during rollback process, we should continue to select disk for the replica locally on the server instead of semi-auto logic, i.e. relying on the /AdminConfigs/ReplicaAddition Znode to get the disk location.

To distinguish the rollback scenario from regular semi-auto, we will keep track of list of resources migrating to Full-Auto in /AdminConfigs/FullAutoMigration Znode. We will read this list on servers and choose disk locally for resources present in the list during rollback.