Fix a logical error in disk failure handling (#2516)
Before this PR, we removed the failed replicas in helix clustermap, and then remove these replicas from storage manager inmemory maps. This is not working with replication and stats manager. Replication manager and stats manager didn't get notified by this change. This is fine when helix decides to move these replicas to a different host, but it will cause some issues when helix decides to allocate some of these replicas in the same host. Since replication manager still keeps the remote token in manager for these replicas, we would not be able to copy existing data for replicas.

There are two different ways to deal with this issue.

Remove these replicas from replication and stats manager in the disk failure handling method.
Rely on storage manager's state transition callback to deal with replication and stats manager.
I prefer the second approach since we already have all the logic in the callback. We don't have to duplicate the same logic in a different method.