FULL_AUTO: use the old way to deal with disk failures, also remove replicas from replication manager (#2520)
This PR reverted some of the logic in this PR #2516.

We will be doing all the operations in disk failure handler, and also remove replicas from replication and stats manager.
The reason to not rely on the partition state change listener is because when a disk already failed before server starts, all the replicas will be errored out when going through upward state transition. In this case, when we reset partitions, those replicas would be transitioned to offline and then to bootstrap, not dropped, if helix decided to keep those replicas in the same host. If we don't remove those replicas from property store, then those replicas would not be treated as bootstrap replicas, and server won't find a new disk for them. That's why we have to remove them from property store. So we do all the operations in the disk failure handler.