Add configuration to enable disk failure handler and fix a replication bug (#2524)
Add a new configuration to enable disk failure handler and fix a replication bug.

When we disable the replica, it would also store a "disabled" information in replica thread to stop replication for this replica. But this is buggy, since when helix decided to add this replica back to this host, this piece of information still exist and this replica can't replicate from other peers.

This pr would fix this issue.