[Efficient Metadata Operations] Remove warn log for partition class not found during metadata reservation (#2526)
The partition class of a partition is determined by the container properties. But its possible that a cluster doesn't have the partition class that is specified in the container properties. In such cases, we just use whatever partition class is present in the cluster. In order to avoid excessive logging for such containers, this PR converts a log message from warn to debug.