Resume decommission when replica state is not inactive (#2551)
When zookeeper session expires, helix would retry to reconnect to zookeeper and then reset the state for all partitions. After reset, most of the partitions would be promoted to BOOTSTRAP -> STANDBY/LEADER, but some of the partitions would be transitioned to DROPPED. In this way, we have to resume the decommission for those partitions since they didn't go through standby -> inactive -> offline callback.