Simplify discarding of header bytes in http2 server code (#2567)
Move code for discarding first 8 bytes of ambry server requests from NettyServerRequestResponseChannel.java to AmbryNetworkRequestHandler.java to avoid calling the code in multiple places.

(Background on these 8 bytes are: When sending requests from frontend to server using Socket server stack, we use first 8 bytes to represent the size of the content so that we know how many bytes to be read from socket. This size header is not needed for HTTP2 server stack since Ambry requests go as content in HTTP2 and Netty stack reads the exact content. To maintain backward compatibility, we just discard first 8 bytes of the request in HTTP2 server stack.)