[vcr-2.0] Fix updateTTL and logging issues (#2611)
This patch fixes updateTTL issues from prev backup code and a few logging issues detailed below.

Replication: When repl-layer receives PUT+TTL together, it applies PUT to store a blob with expiry = -1 followed by updating its expiry to -1. This is redundant and was caught by the new backup code. It floods the logs with unnecessary messages. We can't touch repl-layer at this point so we just suppress the logs.
Backup code V1: The previous backup code reported a blob was not ttl-updated even though it was. This sent duplicate writes to Azure and was caught by the new backup code. This is not only wrong but also floods the logs with unnecessary messages.
Azure SDK: The SDK prints out unnecessary warnings when searching for an absent blob in Azure Storage. The issue is confirmed by MSFT and they'll fix it in their next release. Additionally, we also print some unnecessary messages when we encounter blob_not_found. This patch eliminates those as well.
MSFT issue: Azure/azure-sdk-for-java#37117