[vcr-2.0] Remove concurrent update errors
As VCR catches up with all replicas and approaches current time,
it starts getting the same updates from all replicas. Multiple VCR
threads send the same request to cloud and hit concurrent-update error.
The metrics emitted and log messages printed are not really useful as
they do not indicate a real issue. This patch removes those metrics
and converts the log msgs to a trace msgs.