[vcr-2.0] Remove redundant metadata
The previous backup code 1.0 is a bit weird.

It uses two ways to identify a permanent blob.
1. A lack of expiration-time means the blob is permanent since it was PUT
2. expiration = -1 also means the blob is permanent, after TTL-UPDATE

This distinction is not useful and not used at all during recovery.

Similarly, a lack of deleteTime means the blob has not undergone
deletion and must not be purged from cloud.
A deleteTime = -1 means the blob was undeleted and must not be purged
from cloud.
It would be clearer if we just remove the deleteTime field from metadata
to indicate the blob must not be purged. This granularity of info
is not useful and complicates compaction code with lots of if-else
cases.