Start ReplicationManager with empty replicas (#2640)
Currently ReplicationManager is not started correctly (i.e. started variable is not set and ReplicaTokenPersistor thread is not scheduled) if there are no replicas assigned to the host at boot up. In FullAuto, a new node joining the cluster always has 0 replicas assigned to it initially.

To fix this, we need to remove below condition in ReplicationManager.start() method.

if (replicaThreadPoolByDc.size() == 0) {
        logger.warn("Number of data centers to replicate from is 0, not starting any replica threads");
        return;
      }